Moved SimpleGeoSphereRenderer to mmstd_moldyn
