+++ plugins/mmstd_moldyn/Shaders/simplegeosphere.btf
